Update Sentry.Samples.GenericHost (#1252)
